vpn-idp-client/requirements.txt
websockets
flask-sock
flask
requests
python-dotenv